requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
attrs==25.3.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.0
flet==0.27.6
flet-cli==0.27.6
flet-desktop==0.27.6
flet-web==0.27.6
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
narwhals==1.27.1
numpy==2.2.3
oauthlib==3.2.2
openai==1.68.2
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
propcache==0.3.0
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyLaTeX==1.4.2
PyMsgBox==1.0.9
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
pytweening==1.2.0
pywhatkit==5.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
sympy==1.13.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
unicorn==2.1.3
urllib3==2.3.0
uvicorn==0.34.0
watchdog==4.0.2
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.18.3
